requirements.txt
accelerate==1.12.0
aiobotocore==2.25.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aioitertools==0.13.0
aiolimiter==1.2.1
aiosignal==1.4.0
anndata==0.12.6
annotated-types==0.7.0
anthropic==0.75.0
anyio==3.7.1
array-api-compat==1.12.0
asttokens==3.0.1
attrs==25.4.0
beautifulsoup4==4.14.2
biopython==1.86
biothings_client==0.4.1
bitarray==3.8.0
blinker==1.9.0
bm25s==0.2.14
botocore==1.40.70
bs4==0.0.2
cellxgene-census==1.17.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colbert-ai==0.2.22
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
dataclasses==0.6
dataclasses-json==0.6.7
datasets==4.4.1
decorator==5.2.1
dill==0.4.0
distro==1.9.0
docstring_parser==0.17.0
donfig==0.8.1.post1
et_xmlfile==2.0.0
evaluate==0.4.6
executing==2.2.1
filelock==3.19.1
Flask==3.1.2
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
fuzzywuzzy==0.18.0
gdown==5.2.0
gget==0.29.3
gitdb==4.0.12
GitPython==3.1.45
google-crc32c==1.7.1
greenlet==3.2.4
gritlm==1.0.2
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ipython==9.7.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
langchain==0.0.316
langchain-core==1.1.0
langchain-text-splitters==1.0.0
langdetect==1.0.9
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.0.92
legacy-api-wrap==1.5
lightning-utilities==0.15.2
littleutils==0.2.4
llm2vec==0.2.3
llvmlite==0.45.1
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
mteb==2.2.2
multidict==6.7.0
multiprocess==0.70.18
mygene==3.2.2
mypy_extensions==1.1.0
mysql-connector-python==9.5.0
natsort==8.4.0
networkx==3.5
ninja==1.13.0
nltk==3.9.2
numba==0.62.1
numcodecs==0.16.5
numpy==1.26.4
nvidia-cublas==13.0.0.19
nvidia-cuda-cupti==13.0.48
nvidia-cuda-nvrtc==13.0.48
nvidia-cuda-runtime==13.0.48
nvidia-cudnn-cu13==9.13.0.50
nvidia-cufft==12.0.0.15
nvidia-cufile==1.15.0.42
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.3.29
nvidia-cusparse==12.6.2.49
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.27.7
nvidia-nvjitlink==13.0.39
nvidia-nvshmem-cu13==3.3.24
nvidia-nvtx==13.0.39
ogb==1.3.6
openai==2.8.1
openpyxl==3.1.5
orjson==3.11.4
ormsgpack==1.12.0
outdated==0.2.2
packaging==25.0
pandas==2.3.3
parso==0.8.5
patsy==1.0.2
peft==0.18.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
polars==1.35.2
polars-runtime-32==1.35.2
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyarrow-hotfix==0.7
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pynndescent==0.5.13
pyparsing==3.2.5
PySocks==1.7.1
PyTDC==0.4.17
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytrec_eval-terrier==0.5.10
pytz==2025.2
PyYAML==6.0.3
rdkit==2023.9.6
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
s3fs==2025.9.0
safetensors==0.7.0
scanpy==1.11.5
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
sentence-transformers==5.1.2
sentry-sdk==2.46.0
session-info2==0.2.3
shapely==2.1.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
somacore==1.0.28
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
stark_qa==0.1.2
statsmodels==0.14.5
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiledbsoma==1.17.1
tokenizers==0.19.1
torch==2.9.1+cu130
torch-geometric==2.7.0
torchmetrics==1.8.2
torchvision==0.24.1+cu130
tqdm==4.67.1
traitlets==5.14.3
transformers==4.44.2
triton==3.5.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
umap-learn==0.5.9.post2
urllib3==2.5.0
voyageai==0.3.5
wandb==0.23.0
wcwidth==0.2.14
Werkzeug==3.1.3
widgetsnbextension==4.0.15
wrapt==1.17.3
xxhash==3.6.0
yapf==0.43.0
yarl==1.22.0
zarr==3.1.5
zstandard==0.25.0
